Initial Commit.
A* and Dijkstra work. Use a Priority Queue (Heap). 

A* Currently only works for Nodes, not for Grids. (Dijkstra does not use a Heuristic, thus every grid is the same for it. Memory could however get more performant, as you dont need an Adjacency Matrix and could just calculate the indices on the fly)